rename by_pass_cache
